shanshui-playsave-unity - 1/python/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
moviepy>=1.0.3
keyboard>=0.13.5

